[Add] Move struct FilmCastCard to Reusable Folder as a new file
